✨ (src/app/checkin/*): Add originalEndTime parameter to handle reservation end time updates
This commit adds a new `originalEndTime` parameter to the check-in flow. This parameter is used to detect if the end time of a reservation has been changed. If the end time has been changed, the Google Calendar event is updated accordingly.  This improves the user experience by ensuring that the Google Calendar event always reflects the most up-to-date information.  The update also simplifies the logic for adding events to Google Calendar by removing redundant code.